Load combining
Add load combining that can combine different size of loads such as
   load i8, p; load i8, p+1; load i16, p+2
to
   load i32.
or
   load i16; load i16; load <2xi32>
to
   load <3xi32>

Load combining shares the same core functions with store combining.
They both use either layout struct or vector.

This is the 1st draft and functionality should be complete. By default,
it is off. It will be turned on later once more testing is done.