legalizing gvload users which's further optimizations by standard LLVM passes may result in potential clobbering.
VLoad's value usage was later optimized by mem2reg in a way that resulted in genx_volatile value clobbering.
Current solution to avoid this is to forbid direct usage of genx.vload(@GENX_VOLATILE_GLOBAL*) values by non-genx intrinsic.
Until all the frontends are fixed, we fix-up such cases (emitting warning message in Debug build), since they are trivial to fix-up.

This fix will be followed up by removal of previous fix to this issue which was based on convertion of a non-genx_volatile global
used "in the context of" a genx_volatile global to genx_volatile global. This fix results in implicit conversion of "ordinary globals"
to genx_volatile which is not expected by the user and increases register pressure.