Add alias handling for struct var
When creating cvariable for a struct, the alias checking needs to
be done as well.

This change adds alias checking. It also sets the name of merged load
to have prefix "StructV", so that it helps writig lit tests.

Lit tests will be added after the new llvm15 patch has been propagated into llvm
prebuild.