Fix nullptr deref
When entryBB ends with
call, its RET__loc is identified as LiveThrough since it's defined in
entryBB and used in some other function. This is not really a problem
because RET__loc's interference is correctly captured by SIMT
interference. Whenever we see a call site, in handleCallSite(), we
assume that RET__loc belongs to DefBeforeCallSite so we create an empty
interval for it, which in this special case never gets updated and
interval start/end remains nullptr. When emitting debug info ranges,
these nullptrs are dereferenced which causes a crash. Fix is to rely on
sortedIntervals when updating debug info.