Fix metadata dump: serialize MD in SerializePrintMetaDataPass
When PrintAfter, PrintBefore or ShaderDumpEnableAll is enabled, IR files
dumped before `afterUnification` have outdated metadata, e.g. FuncMD is
empty, due to module metadata is saved in CodeGenContext. Wrong metadata
makes IR file unusable for LIT test of a pass that depends on FuncMD.

This PR serializes metadata to the module in SerializePrintMetaDataPass
and run the pass unconditionally so that correct metadata is dumped.

Behavior of env IGC_PrintMDBeforeModule is not changed, i.e. metadata is
only dumped before the module only if this env is set.