Fix non-deterministic compilation
Weak edges were stored in a set that caused difference in iteration
order of contained elements. During graph relaxation, this different
iteration order caused a secondary effect. Relaxing a node may make its
neighbor unconstrained, which causes that neighbor to be pushed to a
list immediately. When weak edges are iterated in different order, it causes
neighbors to become unconstrained in a different order too that makes
unconstrained list different, and hence changes final coloring order.

Fix is to use a container that behaves like a set so it contains unique
items only and also provides stable iteration order.