Fix init of `IGC_StackOverflowDetection` debug flag
For the purposes of `IGC_StackOverflowDetection` we need to save base
stack pointer (without an offset) in `__stackoverflow_init` function.
In some cases we were adding an offset to the stack pointer (to reserve
data for private memory variables) before calling `__stackoverflow_init`.
This lead to some edge cases of stack-overflowing not being properly
detected. This commit inserts `__stackoverflow_init` directly after
first initialization of `SP_0`. It also removes old calls to
`__stackoverflow_init` and `__stackoverflow_detection` that have to be
present in the kernel for most of the compilation so they don't get
eliminated as dead code.