Optimize DivergentBarrierWA uniform loads
DivergentBarrierWA emits load/store instructions across functions, the complexity
makes it very difficult for WIAnalysis to determine the dependency of the load instruction,
so it will default to RANDOM, which creates high register pressure and spilling.

This optimization runs WIAnalysis on the store instructions, and since the loads will have
a corresponding slot with the stores, we can force uniform loads via WaveShuffleIndex if only
uniform values are stored to that slot.