[Autobackout][FuncReg]Revert of change: 932be31608f3d508a56656625d1bd383d378097a  Restrict augmentation live-intervals to "home" function only
Consider following program:
main:
a =
call foo
= a
b =
call foo
= b

foo:
c = 1
return

a, b are both live in foo due to IPA. Howver, their liveness is
temporally disjoint so they shouldn't be marked as interfering. However,
a and c as well as b and c must be marked as interfering.

As per logic prior to this PR, 2 sub-intervals will be created for a
(b). First live-interval extends from definition of a (b) to use and
second on exists from start to end of foo() given that a (b) is
live-in/live-out of foo().

This PR introduces concept of "home" function. Each G4_Declare has a
"home" function that's either nullptr or pointer to an actual FuncInfo*
instance. When a G4_Declare is defined or used in a single function, its
"home" function is set to that function's FuncInfo instance. Whereas
when a G4_Declare is referenced in multiple functions (eg, kernel and sub),
its "home" function is not unique and is set to nullptr.

In augmentation, live-interval construction for a variable depends on its
"home" function. If "home" function is set to nullptr (ie, args, retval) then
live-intervals are constructed across functions. This means one sub-interval
may be in one function and another sub-interval may be in another function.
Whereas when "home" location contains a non-nullptr value, we restrict
live-intervals to only be contained in "home" function. It means that in above
example, variables a and b have their home location set to kernel and a single
non-overlapping live-interval is created for each. Since no live-interval is
created for these in function foo(), they don't overlap and therefore there's
no interference between them.