Relax assertion
Previously mergeUseFrom() requires both arguments to be used and their
LVInfo have been constructed when mergeUseFrom() is invoked. This isn't
true for inline asm call, in which the second argument is dead in term
of llvm value usage (but call isn't dead). Relax it to let this case
compile without assertion.