Generate an explicit copy for pointers with `byval` attribute
IGC relied on the fact that frontend compilers will generate an
explicit copy (alloca + memcpy), in a caller, for `byval` attributes.
But it turns out that `byval` attribute itself implies that `"a hidden
copy of the pointee is made between the caller and the callee"`.
The "hidden copy" means that a backend compiler should generate a copy,
for all arguments with `byval` attribute. Therefore frontend compilers
are not obligated to generate an explicit copy if they generate `byval`
attribute. Either one or another.
This change adds generating an explicit copy for arguments with `byval`
attribute. So far, adding it under `EnableExplicitCopyForByVal` flag, as it
may cause performance regressions in the following cases:
1) If a frontend compiler generated an explicit copy and `byval` attribute
   simultaneously, it would result in generating a second copy by IGC.
2) If a kernel did not really needed a copy, as it doesn't write any data
   into `byval` argument in a callee, simultaniously not reading that argument
   later in a caller.