Improve store combining
Two stores are candidate if they are not aliased to each other.
Previously, alias analysis is used to check this. As alias analysis
isn't perfect, it might miss some cases in which two stores have the
constant difference but alias analysis conservatively assumes they
are aliased.

For stores that share the same base with the constant offset between
them, this change will directly check if ranges of two stores overlap
without relying on alias anlysis, as this gives the precise result.