Lower `memcpy` right after `LowerByValAttribute` pass
`LowerByValAttribute` inserts calls to memcpy intrinsic. This change introduces
calling `ReplaceUnsupportedIntrinsics` pass right after `LowerByValAttribute`
to transform all `memcpy` calls to a series a loads/stores. Without that,
`LegalizationPass` doesn't legalize all loads and stores, as they are in a form
of `memcpy` calls.

This change also implements the possibility to run `LowerByValAttribute` pass
only if optimizations are enabled. `LowerByValAttribute` pass generates an
explicit copy (alloca+memcpy) for all function arguments with `byval` attribute.
The copy is not necessary if a SPIRV was compiled in O0, as in such case,
FE compilers will generate an explicit copy (not optimize it out) in SPIRV.