Move loaded constants outside of loops
GenXCategory pass loads constants and produced intructions are placed
just before their users. If a user is inside a loop, then it results in
immediate MOVs inside loop's body in vISA, because there is no LICM pass
after GenXCategory.
Loaded constants have to be moved just before the loop to avoid that