Fix bindless loads merging on OpenCL path
When `ConstantCoalescing` pass merges 3 scalar loads into one, it needs to
decide whether to produce 3 or 4 element vector load based on what send
message is expected to be produced from that instruction in `EmitVISAPass`.

If a bindless load is expected to be transformed into LSC message, then
`ConstantCoalescing` can produce 3 element vector load, because LSC messages
support 3 element vectors.

But if a bindless load is expected to be transformed into legacy message, then
`ConstantCoalescing` must produce 4 elements vector load (round vector size up),
because legacy messages don't support 3 element vectors.

Without this change, `ConstantCoalescing` only checks if a target platform
supports LSC, but it doesn't check if LSC message is available for the bindless
load instruction it aims to produce. This fix replaces usage of `platform.LSCEnabled()`
with `shouldGenerateLSCQuery` which takes a load instruction and verifies if a
LSC message can be produced for that particular instruction.

This is crucial, because LSC messages are currently disabled for bindless
loads and stores, so it's necessary to produce 4 element vector load, so that
it can be properly handled by a legacy message.