Enable stateful promotion for non-gep pointers
For some unknown historic reasons, stateful promotion was disabled for
loads and stores that use non-gep pointers, meaning that the pointer used
by load/store instruction is not a product of `getelementptr` instruction,
but it comes (for example) directly from a kernel argument without any offset.

This change affects both: bindless and bindful promotion.