Fixed function cloning check when caller is an indirect call
This is a bug in GenXCodeGenModule::detectUnpromotableFunctions() when detecting functions promotable
to indirect call. If a function is not promotable, we have to mark all its callers as unpromotable as well.
However, when recursively checking callers, we have to check non CallInst users as well, in case it's an
indirect call.