Fix detection of generic pointer for `memcpy`, `memmove` and `memset`
`CheckInstrTypes` pass detects load and store instruction that operate on
a generic pointer. If any of them is present, then IGC allocates private
variables in global memory, hence avoids usage of scratch space for private
memory, to avoid the necessity for generating if/else control flow to handle
generic pointer memory operations.

It may happen that neither raw load nor raw store is present in a kernel,
but `memcpy`, `memmove` or `memset` intrinsics that operate on a generic
pointer are present. These intrinsics will be lowered into a series of
raw loads and stores at some compilation point, so they should be detected
by `CheckInstrTypes` pass and `hasDynamicGenericLoadStore` should be set.