New Block2D APIs that take constructed address payload
For performance, address payload construction for block2D messages are
handled by IGC rather than vISA during construction of block2D send
message. This is useful for benchmarks with loops containing block2d
instructions. The address payload is loop invariant and it is better to
hoist this payload construction outside of the loop.