Fix denorm flush-to-zero issue in math macro. 4th try.
Single Precision or Double precision denorm mode in control register must be set
to retain denorm mode when executing Math Macro instruction sequence. Besides, if
vISA_hasRNEandDenorm is not set, the denorm mode could be flush-to-zero. When
denorm flush-to-zero is set, mov instructions with the same source and destination
data type may retain denorm as output. So, we need to insert a mul instruction with
1 to make denorm flush-to-zero take effect.