Add pseudo-kill when filling entire address taken variable
When an address taken variable is indirectly referenced, the compiler
cannot statically determine which parts of the variable are being
referenced. Therefore, it has to assume any part of the variable may be
read/written.

In case such a variable is spilled, we need to bring entire variable to
registers in order to read/write to it. In some cases, aftering filling
entire variable in a temporary, spill cleanup may optimize away some
rows of this temporary because those rows may have never been defined.
When this happens, subsequent liveness pass extends the filled
temporary liveness above what is required.

We insert a pseudo-kill instruction just before filling in the temporary
so that even if some rows are eliminated, liveness pass models KILL
location appropriately.