Workaround __builtin_IB_get_address_mode for indirect sampler
__builtin_IB_get_address_mode, __builtin_IB_get_snap_wa_reqd and
__builtin_IB_is_normalized_coords are used in sampled image builtins.
For indirect sampler, the latter two builtins are workarounded in
ImageFuncResolution pass by replacing with 0. The first builtin should
be workarounded as well.
This PR is also a workaround for SYCL bindless sampled image since we
are not able to resolve these three builtins currently for sampler
loaded from memory. This PR fixes compile fail in test read_sampled.cpp
and read_write_user_type.cpp.