[Autobackout][FuncReg]Revert of change: a0caa5f28bb7674ce6f3d22ace0112932668d4c5  Using ptr as address payload's type
By design, address payload updating builtins does not create a new
address payload.  If the payload's type is int, not pointer, the
updating builtin would be like the following:
    (1) int addrP0 = createAddrPayload(...)
    (2) v = block2d_read (addrP0, ....)
    (3) int addrP1 = setBlockX(addrP0, ...)
In llvm IR, those addrP0 and addrP1 are different llvm values. And
it is legal for llvm to reorder instructions into { (1), (3), (2) },
which is incorrect as (2) should use the original addrP0, not the
updated one (addrP1).

For this reason, address payload should be of pointer type. With ptr
type, the above would be:
    (1) int* addrP0 = createAddrPayload(...)
    (2) v = block2d_read (addrP0, ....)
    (3) etBlockX(addrP0, ...)

The llvm cannot reorder them as there are dependences b/w (2) and (3).

This PR makes address payload's type to be pointer type (int*).