Fix non-deterministic code generation
std::map with ptr as key has non-deterministic order of iterating
its members as ptr would be different from run to run. This non-
deterministic order causes visa to generate different code.

This PR fixes it by sorting map using a fixed value id. With this,
it will generate the same order of cvariables for vector alias for
the same input function.