Account for negative indices when resolving GEP insts.
TransposeHelper::handleGEPInst creates nuw add regardless of what
operand is supplied to the gep instruction handled. That can be
detrimental in case of negative operands, 'add nuw x, -1' being an
example, which gets optimized to just the value of -1, since any value
of x other than 0 would mean an overflow, which can't happen because of
the 'nuw' tag. This commit fixes that by checking each operand for
negativity and creating non-nuw add insts accordingly instead.