Discard value names in BiF modules
The value names inside built-in function modules are not needed for
the user's code generation. However, they are still present in the IR,
which increases the size of the bitcode twice. This change removes the
value names from the BiF modules to reduce the size of the IGC shared
library.