Fixed bug with Inline Asm vector copy
When there are multiple outputs for an inline asm block, the output is represented as a struct type.
This change fixes a bug in EmitVISA for inline asm instructions, where we check the struct output instead of individual struct elements when determining if we need to do a src to dst copy.