Prefer adding all offsets first
Description:
This PR is to revert the previous change

Previous change:
   threadBase = privateBase + threadOffset

   loop:
       addr = threadBase + simdBufferOffset

With this change, it becomes:
   loop:
       tmp = threadOff + simdBufferOffset
       addr = private + tmp

Since tmp is a loop invariant and will be hoisted out of the loop, both
will only have addr inside the loop.

If the total private memory for one kernel is less than 4GB,
the previous one has two 64bit adds, where the new one has one 32bit
add and one 64bit add. On platforms without native 64bit support, this