Improve private memory SOA transpose
Improve private memory SOA transpose by allowing array of struct to
be transposed. If a struct is a simple one, it is divided into
chunks of the fixed size and each chunk is transposed.

The new algorithm can handle both dw[xn], qw[xn], array of struct, and
it is super set of the existing one. The plan is to replace the old
one with the new one.

This change has all code and lit tests, but it is off by default.
The plan is:
   (1) turn it on for array of struct, where struct is simple
   (2) (1) + qw[xn], dw[xn].  This will be able to replace the existing one.
   (3) array of struct, where struct is more complicated.