Prepare for turning on new PM SOA transpose
Prepare turning on new PM SOA transpose optimization by
modifying lit test flag.

Also, the existing code does not promote (to register) array of i8 if the array
is too big. This change advoid promoting array of i8 vector as it should be treated
the same as array of i8.