Generalize LoopSplitWidePHIs pass for more concat idioms
Extends the functionality of the LoopSplitWidePHIs pass to split PHI
nodes into >= 2 parts, when the PHI loop value is effectively a
catenation of parts and all uses of the PHI extract one of those parts.
Besides allowing for concat/extract shuffles to be eliminated, the pass
now eliminates any intervening bitcasts or identity-with-padding
shuffles present in the concat/extract idioms.