Fix issue related with GET_MEMPOOL_PTR
The macro GET_MEMPOOL_PTR in IGC compiler allocs the SLM memory for any usage.
The allocation is made only once per Function. All GET_MEMPOOL_PTR calls (in Function) are collect as request for the SLM.
From the all requests the largest one defines the size of the SLM allocation, which then will be shared between
all built-ins functions which was requesting for this memory spot. We need to insure that any access to this common SLM
memory chunk will not be overwrite between the built-in functions - with this wg_scope barrier inside GET_MEMPOOL_PTR
we could prevent any race-access to common SLM.