Avoid stack overflow in `BasicAA` by adjusting the recursion limit
Patch-in an LLVM option to adjust the maximum query depth within BasicAA, and
set the limit to a more conservative value. This minimizes the risk of stack
overflows for particularly heavy cases of loop unrolling. Fundamentally, this
is caused by suboptimal implementation of the iteration logic itself, and the
original limit has long existed as a workaround for that same issue in the
community.

Co-authored-by: Mateusz Chudyk <mateusz.chudyk@intel.com>