Extend support for creating SCEVs iteratively
This change is to avoid stack overflow issues related to creating scalar evolution expressions. The previous solution can lead to unintendent scenarios when building an expression chain requires to use a recursive approach extensively. Moreover, there is an improvement in the quality of these expressions because of creating a final expression of the phi instruction only in the recursive function.