Replace genx.any predicate operation with compare
With the GenXPromotePredicate pass produsing scalar bitwise operations
on predicates, the genx.any operation can generate inefficient code.
So, it's profitable to replace the genx.any intrinsic with a compare
instruction.

The following pattern is matched:
```LLVM
%1 = bitcast i16 %0 to <16 x i1>
%2 = call i1 @llvm.genx.any.v16i1(<16 x i1> %1)
```

The matched sequence is replaced with the `icmp` instruction as follows:
```LLVM
%1 = icmp ne i16 %0, 0
```

For `icmp` instruction we don't need to move the value from GRF to
a flag register, so it's more efficient.