Adjust BFN matching heuristics for pattern matching
VC shouldn't match some non-profitable cases of BFN patterns:
* ARF is not supported as a source for BFN
* Flag register is not supported as a source for BFN

The finalizer produces extra MOV instructions in these cases, so it's
better to avoid them.

We also need to reduce the use threshold for instructions folded into
the BFN. This will reduce the value live ranges and improve the
register pressure.