Lower dpas accumulator for bfloat16 in VC
The dpas intrinsics accumulator source and return value can be bfloat16.
The `i16` data type can be used as a storage. However, the finalizer
requires the accumulator to be `bf` register in such case. If VC
translates `i16` into `bfloat` for the accumulator source and the return
value, we can simplify the CisaBuilder.