Fix for reduction
For some work-group sizes the current implementation of reduction is not working well.
The implementation in repo assumes that all of work-items will be executed.
Even if the workgroup size isn't the multiple of the width of SIMD size.
This change returns the SLM+barrier for the final calculation of the reduction (performance degradation).
TODO: Remove the SLM+barrier and force execution of the whole reduction built-in function with NoMask on asm level.

(cherry picked from commit c147d7f520f2a021663d913f69438bbe0366c464)