Adjust BFN matching heuristics for pattern matching
VC shouldn't match non-profitable patterns. Flag register cannot be
source for BFN instruction. The finalizer emits extra MOV instructions
in such cases. So, it's better to avoid them, when the total amount of
instruction is not reduced.

We also need to reduce the use threshold for instructions folded into
the BFN. This will reduce the value live ranges and improve the
register pressure.