Improvement to logic deciding between block and scatter spill
If an instruction has a source operand that fully overlaps with
destination and if the underlying variable was spilled, it means that a
block fill will always be emitted. Even when the instruction doesn't use
NoMask, we can use block spill for such cases instead of using scatter.
Block spill can get coalesced with nearby spills whereas scatter spills
cannot get coalesced.