Add pre-header to main subroutine's FuncInfo instance
VISA creates new pre-header BB for a loop if it cannot find one. When a
new pre-header is created, it has to be added to FuncInfo* instance of
either kernel or the subroutine.

The problem was when there's no subroutine, sortedKernelFunc vector is
empty. So we need to use special logic to handle this scenario.

Assert is also added to catch the case where a pre-header BB is created
and added to CFG but not added to any FuncInfo*.