Pass to convert 2D block intrinsics into constituents.
The 2D block IO intrinsics are opaque to the compiler and identical
payloads cannot be optimized by the compiler. Intrinsics have been
introduced to overcome this problem by allowing the 2D intrinsics to be
broken up into constituent stages (payload, set X and Y relative to
payload, IO operation).

This PR adds a new pass, `Decompose2DBlockFuncs`, which converts the 2D
IO intrinsics into the constituents, allowing subequent passes to
optimize the code. At the current stage, the payload intrinsics can be
hoisted above loops by LICM, and can be merged by EarlyCSA and others.

The `LSC2DBlockSetAddrPayloadField` intrinsics have a mode in which the
IO location in memory is specified exactly (relative to the payload),
and another mode in which the intrinsic increments the value. When a
payload is used by the `LSC2DBlockSetAddrPayloadField` intrinsic in the
accumulator mode, the payload cannot be hoisted. Because of this, I have
given the `LSC2DBlockCreateAddrPayload` default attributes of
`writeonly`, which is conservative for when the SetAddr intrinsic is in
accumulator mode.

The `LSC2DBlockSetAddrPayloadField` intrinsic added by the
`Decompose2DBlockFuncs` is always in the mode in which it explicitly
specifies the memory location. Since we have this guarantee, the related
payload is described as not accessing memory, since the dependencies are
all properly accounted for by the data dependency; this allows the
desired optimizations to function.

The 2D block intrinsics are not decomposed when they do not fall within
a loop, or if the payload would have loop-dependent parameters.