Fixing selective scalarization
ScalarizeFunction pass can keep some instructions vectorized, if the vector is used as the whole entity. The pass builds a web of instructions protected from scalarization. The ending legs of the web consist of vectorial instructions such as insert and extract elements, vector shuffles, GenISA intrinsics and function calls. The vectorial instructions inside the web consist of bitcasts and PHI nodes.