Expanding Decompose2DBlockFuncs hoisting opportunities
2D block decomposition in the Decompose2DBlockFuncs pass was not performed in some test cases because the LLVM isLoopInvariant() checks only that the args of the instruction are outside the loop. Moving the pass after an LICM call makes this correct.

Payload hoisting was not performed in some cases because of interfering intrinsics that disrupted the implicit control flow. We can assume, however, payload is speculatable when we are sure we are never changing it in the loop body, fixing this issue.

I also added an IGC flag for easier testing of this feature.