Fix uniformity analysis for `GenISA_WaveBroadcast`
`sub_group_broadcast` f unction(which is implemented by `GenISA_WaveBroadcast`)
is restricted to be executed with uniform local ID argument. IGC utilizes this
information to generate more performant code. However, if a local ID comes
directly from the load instruction, we end up with the load instruction that
has a uniform destination and a non-uniform source, which is the case that is
not supported by `EmitVISAPass`.

(cherry picked from commit 5881894b7952ae75e458fd39417781a8912752d6)