Ignore L1 hit cache option when scheduling
Add -ignoreL1Hit visa flag to ignore L1 hit when calculating
latency for scheduling.  The default is false, meaning not ignoring
cache option.

This is an experimental flag to make sure that scheduling does not
change for different cache options so that cache option effect can
be better understood.

No functional change.