Legalize vector `fneg` instructions
IGC's legalization phase was not able to properly legalize `fneg` instructions
that operate on a vector type. This issue has been exposed while merging
commit f8f17a0cb7fb9ca2e13c11e6fbf13b0db157e2a3 which has disabled scalarization
of vector `fneg` instructions in `ScalarizeFunction` pass for `optnone` functions.

`ScalarizeFunction` pass is not intented to be a legalization pass. It's rather
treated as an optimization pass, so whatever gets disabled in its implementation,
shouldn't affect the final kernel correctness.

`ScalarizerCodeGen` pass is a pass that runs during legalization phase and it's
specifically designed to scalarize instructions that cannot be properly handled
by `EmitVISAPass` if they are in a vector form. This change extends the
capabilities of `ScalarizerCodeGen`, so that it can properly scalarize `fneg`
vector instructions.

This change also fixes a bug introduced in f8f17a0cb7fb9ca2e13c11e6fbf13b0db157e2a3
which sometimes resulted in creating fneg instruction with `undef`` values.