AlignmentAnalysis: set correct dst alignment for memcpy from global constant
If memcpy source is zeroinitialized constant global, memcpy is
equivalent to memset to zero. In this case alignment can be set based
only on destination.

This change also removes early run of LLVM's MemCpy Optimization pass
when Khronos SPIR-V translator is used, as there is no longer need to
change memcpy to memsets before AlignmentAnalysis.