Resolve assertions with Opaque Pointers in Joint Matrix resolution pass
Check if pointer is opaque and do not call get NonOpaquePtrEltTy if so to avoid assertions. The pass will not properly resolve Joint Matrix in case opaque pointers are used, but at least it will not fail with assertions and will allow compilation to go forward.
Add isAnyOperand helper to check types iterating over all operands
Fixed bug: when function signature is resolved, need to also resolve all instructions in function body, which are impacted by change of type in function signature. Otherwise, invalid IR is produced.
Fixed bug: When searching for entry function, parent function was assigned as entry, which is incorrect.
[NFC]: More lit tests to cover fixes for above bugs
[NFC] Refactoring: simplified code in several places.
[NFC] Improve debuggability: update IGC_ASSERT in some places to provide more details, add LLVM_DEBUG support to JointMatrixFuncsResolutionPass
[NFC] fix copyright year that was updated by mistake previously
[NFC] moved some static helpers up as they are used in more methods now and to avoid forward declarations