Fix negative indices handling of getelementptr instructions.
This patch makes the same fix as the previous
change in LowerGEPForPrivMem.cpp. The problem with
the previous patch is that it assumed that it's
checking "each operand" where is actually checking
the "current offset", which can be LLVM
instruction computing the offset. See the
regression test added to the patch.

In general, getelementptr instruction indeces are
signed integers, so applying nuw flag to the
instructions that compute the offset is not
correct. I've added a TODO comment about adding
support for getelementptr attributes.