Generating block write instructions with correct alignment.
The i32 block store instruction is converted to store.ugm.d64x16t.a64, which requires an 8-byte alignment.
The scatter i32 store instruction requires 4 byte alignment. Therefore, in order to replace the scatter with
a block one, it is necessary to generate it with an alignment of 4 bytes. This patch marks the block instruction
with an attribute to later create store.ugm.d32x32t.a64.