Add (disabled) immediate offset capability of 2D block intrinsics
This change adds the capability to enable the use of automatic immediate
offset for 2d block intrinsics. Currently, the pass requires the env var
`allowImmOff2DBlockFuncs` to be turned on manually, since I have not yet done
adequate testing, but enabling the feature is now trivial.

I also refactored the code, to make future optimizations easier to
implement; it is now easy to introduce new set intrinsics for other
parameters of the payload (not enabled now by default), and should be
easier to add the "increment" mode functionality of the
LSC2DBlockSetAddrPayloadField intrinsics.

This change should not affect compiler behaviour. Once one of the
features that this change facilitates is found to be advantageous, it
can be enabled.