Add option to specify file for SelectiveFunctionControl
SelectiveFunctionControl allows user to choose whether a function must
be inlined, emitted as subroutine, or as stack call. It used a default
filename FunctionDebug.txt in IGC output folder. This makes it difficult
to write tests for debugger once we add support for subroutine debug
info.

With this change, user can use SelectiveFunctionControlFile to specify a
custom file with path per test. When this is not specified, IGC falls
back to old behavior of using FunctionDebug.txt from IGC output folder.