GEP LSR pass: don't reduce if calculated address is not used on first iteration
If it can't be proven that calculated address in loop is used on
loop's first iteration, address calculation can't be reduced to
preheader, as there is no guarantee calculated address is valid.
Previously GEP LSR pass only validated dominance of GEP instruction;
this commit extends the check to GEP's users too.