[LLVM16] Port intrinsic definitions to `MemoryEffects` syntax
As a follow-up on commit 1d2335c, this change aims to utilize the new
`MemoryEffects` wrapper infrastructure so that compatible intrinsic definitions
could be generated with legacy memory attributes (LLVM pre-16), as well as the
unified `memory` attribute (LLVM 16). Direct references to legacy memory
attributes are therefore removed from YAML grammars.

The full capabilities of the unified attribute and `llvm::MemoryEffects`, such
as multiple location-specific modifiers and "memory effects arithmetics" like
`llvm::MemoryEffects::operator&`, can only be supported after a full switch to
LLVM 16. Therefore, the implementation is currently restricted to a single
optional specification of a location/access restriction.