Add better support for unnamed pass dumps
Name of the pass dump is based on the return value of getPassName function.
If this function is not overridden, a default value is returned i.e.
"Unnamed pass: implement Pass::getPassName()". A colon is an invalid
character for Windows filenames which causes problems when generating dumps
on Windows or copying dumps generated on Ubuntu.