Improvement in removing noop stores in DSE
This change is to add a patch to llvm addressing issues related to undiscovered noop stores in DSE. Note that noop stores are such store instructions which writes the same values which resides in the memory in this time. MemorySSA produces a new memory state each time the memory is updated. However, finding clobbering states, which aliases/updates/changes this memory state, with the memory access describing the value operand considering optimizations by the walker must be identical. This is missing in the current implementation.