Match and fold whole region override chains in VC
When a vector is fully overwritten by a series of wrregion instructions,
VC can break the false data dependency chain by replacing the initial
vector value with `undef`.

When the vector is overwritten by a series of wrregion instructions, and
then inserted into a larger vector, VC can replace the initial chain
of wrregion instructions with an alternative chain that writes the data
directly into the larger vector.

This transformation eliminates false data dependencies and unnecessary
MOV instructions.