Handle software fence in depressurizer in VC
The GenXDepressurizer pass tries to sink instructions to reduce the
pressure on the register file. It's almost always beneficial, but when
a user explicitly writes a software fence instruction, the pass should
not sink instructions across the fence. This change adds a check to
prevent sinking across the fence.

The software fence instruction is a pseudo instruction that is used as
a hint to the compiler and the finalizer to not reorder instructions
across it during optimizations and scheduling.

The fence can be represented in LLVM IR as either an intrinsic or
an inline assembly instruction as follows:

```LLVM
  call void @llvm.genx.fence(i8 -128)
  call void asm sideeffect "fence_sw", ""()
```

As far as the inline assembly operation can include any string as the
list of instructions, the compiler conservatively assumes that any
inline assembly instruction can be a software fence.