Match and fold whole region override chains in VC
When the vector is overwritten by a series of wrregion instructions, and
then inserted into a larger vector, VC can replace the initial chain
of wrregion instructions with an alternative chain that writes the data
directly into the larger vector.

This transformation eliminates false data dependencies and unnecessary
MOV instructions.