Expand supported operand types in GenericPointersCmpPatterh.
GenericPointersCmpPattern assumes that both operands in a cmp
instruction comparing generic pointers must be results of
ExtractValueInst in case of 64bit emulation (pointers are 64bit wide and
thus emulated to structs of 2 32bit integers). This doesn't always have
to be the case as along the way the ptr value can be selected. It might
also be possible for it to come from other conditional statements. If
that ever happens, more blocks will need to be added to tracksDownToEmu.