Improve transformation of `mul` by power-of-2 constants
This mul-to-shl optimization was initially introduced in commit d2e586f. From a
functional standpoint, this change addresses the fundamental issue with
`nuw`/`nsw` tags being discarded altogether, when we should in fact respect the
tags for all of the newly generated instructions. We should also skip the
transformation for 1/-1, as apparently this can skew some `mad`-based
optimization opportunities in i64 emulation cases.

Tests are adjusted and expanded accordingly. Additionally, early exits are
employed where appropriate, with more detailed TODOs for future improvements.