Handle non-loop backedge in augmentation interval
When there's a non-loop backedge in augmentation interval and any BB in
that interval is a JIP target, it's possible for control to enter in the
middle of that interval. This could cause following sequence of events:
1. Def of a variable is executed,
2. Non-loop backedge is taken,
3. JIP edge that executes the use is taken.

It means that we need to preserve def from (1) till (3) is executed. So
we must extend augmentation interval so that it covers the non-loop
backedge target.