Fix signal handlers for IGC
IGC sets custom handlers for several signals, but the handlers are not
restored when the signal is received. This can lead to a crash if the
signal is received again after the IGC invocation.

This change adds a RAII class to manage the signal handlers. The class
sets the custom handlers when constructed and restores the previous
handlers when destructed.