Break constexprs before SLM resolution
The same constexpr which uses SLM var may be located in different kernels/function groups. This leads to an issue while trying to replace uses of the SLM variable since its user(constexpr) is the same for different kernels which makes impossible to use different offsets for the same SLM var.
The solution is to get rid of constexprs since such issue is not possible with instructions.