Fix indirect pointer detection for structs in VC
If the pointer is the first member of the struct, the compiler would
incorrectly mark it as a byptr argument. The correct way is to mark it
as a byvalue argument with is_ptr set to true.

(cherry picked from commit e572b3b30306d2017687d4a3e4519b5dfc3ae08a)